日志

+++ xiaoerke-web/src/main/java/com/cxqm/xiaoerke/modules/consult/web/BabyCoinController.java
@@ -367,7 +367,7 @@ public class BabyCoinController {
     @ResponseBody
     void giveBabyCoin(@RequestBody Map<String, Object> params,HttpSession session, HttpServletRequest request) {
         String openId = WechatUtil.getOpenId(session, request);//"oogbDwD_2BTQpftPu9QClr-mCw7U"
-        Long count = (Long)params.get("count");
+        Long count = Long.valueOf((Integer) params.get("count"));
         babyCoinService.giveBabyCoin(openId,count);
     }
